seperate full screen video into different component
